requirements.txt
Jinja2
google-cloud-storage